make option for old style timeseries from emitter, pull out Mears plots
